requirements.txt
flask
sklearn 
uvicorn
gunicorn
panda
joblib
numpy